Aggregate health status in unified resource cache
The health status for a resource with multiple servers, as in a database
proxied by multiple agents, will now be aggregated in the unified
resource cache. If there are multiple statuses and the statuses are not
all the same, then the aggregate status will be set to "mixed".
It is then the client's responsibility to list the servers for a "mixed"
status resource if they want details about individual resource server
health.

+++ lib/services/unified_resource.go
@@ -65,7 +65,7 @@ type UnifiedResourceCache struct {
 	typeTree *btree.BTreeG[*item]
 	// resources is a map of all resources currently tracked in the tree
 	// the key is always name/type
-	resources       map[string]resource
+	resources       map[string]resourceCollection
 	initializationC chan struct{}
 	stale           bool
 	once            sync.Once
@@ -97,7 +97,7 @@ func NewUnifiedResourceCache(ctx context.Context, cfg UnifiedResourceCacheConfig
 		typeTree: btree.NewG(cfg.BTreeDegree, func(a, b *item) bool {
 			return a.Less(b)
 		}),
-		resources:       make(map[string]resource),
+		resources:       make(map[string]resourceCollection),
 		initializationC: make(chan struct{}),
 		ResourceGetter:  cfg.ResourceGetter,
 		cache:           lazyCache,
@@ -127,31 +127,28 @@ func (cfg *UnifiedResourceCacheConfig) CheckAndSetDefaults() error {
 func (c *UnifiedResourceCache) putLocked(resource resource) {
 	key := resourceKey(resource)
 	sortKey := makeResourceSortKey(resource)
-	oldResource, exists := c.resources[key]
+	collection, exists := c.resources[key]
 	if exists {
 		// If the resource has changed in such a way that the sort keys
 		// for the nameTree or typeTree change, remove the old entries
 		// from those trees before adding a new one. This can happen
 		// when a node's hostname changes
-		oldSortKey := makeResourceSortKey(oldResource)
+		oldSortKey := makeResourceSortKey(collection.get())
 		if oldSortKey.byName.Compare(sortKey.byName) != 0 {
 			c.deleteSortKey(oldSortKey)
 		}
+		collection.put(resource)
+	} else {
+		collection = newResourceCollection(resource)
+		c.resources[key] = collection
 	}
-	c.resources[key] = resource
 	c.nameTree.ReplaceOrInsert(&item{Key: sortKey.byName, Value: key})
 	c.typeTree.ReplaceOrInsert(&item{Key: sortKey.byType, Value: key})
 }
 
 func putResources[T resource](cache *UnifiedResourceCache, resources []T) {
 	for _, resource := range resources {
-		// generate the unique resource key and add the resource to the resources map
-		key := resourceKey(resource)
-		cache.resources[key] = resource
-
-		sortKey := makeResourceSortKey(resource)
-		cache.nameTree.ReplaceOrInsert(&item{Key: sortKey.byName, Value: key})
-		cache.typeTree.ReplaceOrInsert(&item{Key: sortKey.byType, Value: key})
+		cache.putLocked(resource)
 	}
 }
 
@@ -167,14 +164,16 @@ func (c *UnifiedResourceCache) deleteSortKey(sortKey resourceSortKey) error {
 
 func (c *UnifiedResourceCache) deleteLocked(res types.Resource) error {
 	key := resourceKey(res)
-	resource, exists := c.resources[key]
+	collection, exists := c.resources[key]
 	if !exists {
 		return trace.NotFound("cannot delete resource: key %s not found in unified resource cache", key)
 	}
 
-	sortKey := makeResourceSortKey(resource)
-	c.deleteSortKey(sortKey)
-	delete(c.resources, key)
+	if empty := collection.remove(res); empty {
+		sortKey := makeResourceSortKey(collection.get())
+		c.deleteSortKey(sortKey)
+		delete(c.resources, key)
+	}
 	return nil
 }
 
@@ -243,13 +242,13 @@ func (c *UnifiedResourceCache) iterateItems(ctx context.Context, start string, s
 						}
 					}
 
-					r, ok := cache.resources[item.Value]
+					collection, ok := cache.resources[item.Value]
 					if !ok {
 						return true
 					}
 
-					if len(kinds) == 0 || c.itemKindMatches(r, kindsMap) {
-						items = append(items, iteratedItem{key: item.Key, resource: r})
+					if len(kinds) == 0 || c.itemKindMatches(collection.get(), kindsMap) {
+						items = append(items, iteratedItem{key: item.Key, resource: collection.get()})
 					}
 
 					if len(items) >= defaultPageSize {
@@ -473,7 +472,11 @@ func (c *UnifiedResourceCache) GetUnifiedResourcesByIDs(ctx context.Context, ids
 			if !found || res == nil {
 				continue
 			}
-			resource := cache.resources[res.Value]
+			collection, ok := cache.resources[res.Value]
+			if !ok {
+				continue
+			}
+			resource := collection.get()
 			match, err := matchFn(resource)
 			if err != nil {
 				return trace.Wrap(err)
@@ -660,20 +663,7 @@ func (c *UnifiedResourceCache) getDatabaseServers(ctx context.Context) ([]types.
 	if err != nil {
 		return nil, trace.Wrap(err, "getting database servers for unified resource watcher")
 	}
-	// because it's possible to have multiple replicas of a database server serving the same database
-	// we only want to store one based on its internal database resource
-	unique := map[string]struct{}{}
-	resources := make([]types.DatabaseServer, 0, len(newDbs))
-	for _, dbServer := range newDbs {
-		db := dbServer.GetDatabase()
-		if _, ok := unique[db.GetName()]; ok {
-			continue
-		}
-		unique[db.GetName()] = struct{}{}
-		resources = append(resources, dbServer)
-	}
-
-	return resources, nil
+	return newDbs, nil
 }
 
 // getKubeServers will get all kube servers
@@ -682,18 +672,7 @@ func (c *UnifiedResourceCache) getKubeServers(ctx context.Context) ([]types.Kube
 	if err != nil {
 		return nil, trace.Wrap(err, "getting kube servers for unified resource watcher")
 	}
-	unique := map[string]struct{}{}
-	resources := make([]types.KubeServer, 0, len(newKubes))
-	for _, kubeServer := range newKubes {
-		cluster := kubeServer.GetCluster()
-		if _, ok := unique[cluster.GetName()]; ok {
-			continue
-		}
-		unique[cluster.GetName()] = struct{}{}
-		resources = append(resources, kubeServer)
-	}
-
-	return resources, nil
+	return newKubes, nil
 }
 
 // getAppServers will get all application servers
@@ -702,18 +681,7 @@ func (c *UnifiedResourceCache) getAppServers(ctx context.Context) ([]types.AppSe
 	if err != nil {
 		return nil, trace.Wrap(err, "getting app servers for unified resource watcher")
 	}
-	unique := map[string]struct{}{}
-	resources := make([]types.AppServer, 0, len(newApps))
-	for _, appServer := range newApps {
-		app := appServer.GetApp()
-		if _, ok := unique[app.GetName()]; ok {
-			continue
-		}
-		unique[app.GetName()] = struct{}{}
-		resources = append(resources, appServer)
-	}
-
-	return resources, nil
+	return newApps, nil
 }
 
 // getDesktops will get all windows desktops
@@ -807,7 +775,7 @@ func (c *UnifiedResourceCache) read(ctx context.Context, fn func(cache *UnifiedR
 			typeTree: btree.NewG(c.cfg.BTreeDegree, func(a, b *item) bool {
 				return a.Less(b)
 			}),
-			resources:       make(map[string]resource),
+			resources:       make(map[string]resourceCollection),
 			ResourceGetter:  c.ResourceGetter,
 			initializationC: make(chan struct{}),
 		}
@@ -938,6 +906,152 @@ func (i *item) Less(iother btree.Item) bool {
 type resource interface {
 	types.ResourceWithLabels
 	CloneResource() types.ResourceWithLabels
+}
+
+type resourceCollection interface {
+	get() resource
+	put(r resource)
+	// remove removes a resource from the collection and returns true if the
+	// collection itself should be removed.
+	remove(r types.Resource) bool
+}
+
+func newResourceCollection(r resource) resourceCollection {
+	switch r := r.(type) {
+	case types.DatabaseServer:
+		return newDatabaseServerCollection(r)
+	case types.AppServer, types.KubeServer, types.WindowsDesktop:
+		return &resourceServerCollection{
+			latest:  r,
+			servers: make(map[string]resource),
+		}
+	default:
+		return &singularResourceCollection{latest: r}
+	}
+}
+
+type singularResourceCollection struct {
+	latest resource
+}
+
+func (c *singularResourceCollection) get() resource { return c.latest }
+
+func (c *singularResourceCollection) put(r resource) { c.latest = r }
+
+func (c *singularResourceCollection) remove(types.Resource) bool { return true }
+
+type resourceServerCollection struct {
+	latest  resource
+	servers map[string]resource
+}
+
+func (c *resourceServerCollection) get() resource {
+	return c.latest
+}
+
+func (c *resourceServerCollection) put(r resource) {
+	c.latest = r
+	if h, ok := r.(interface{ GetHostID() string }); ok {
+		c.servers[h.GetHostID()] = r
+	}
+}
+
+func (c *resourceServerCollection) remove(r types.Resource) bool {
+	// This looks insane, but we only get a [types.ResourceHeader] in
+	// [types.OpDelete] events.
+	// The types that actually implement [resourceServer] all store the host ID
+	// in the description of the resource header metadata on deletion.
+	// If a new type is added that implements [resourceServer] and the
+	// unified resource watchers starts watching it, then please:
+	// - add it to the isResourceServer helper func
+	// - ensure host ID is stored in the metadata description in its event parser
+	// - add test coverage for it in TestUnifiedResourceWatcher_DeleteEvent
+	delete(c.servers, r.GetMetadata().Description)
+	for _, s := range c.servers {
+		c.latest = s
+		return false
+	}
+	return true
+}
+
+func newDatabaseServerCollection(db types.DatabaseServer) *databaseServerCollection {
+	statuses := map[string]types.DatabaseServer{
+		db.GetHostID(): db,
+	}
+	return &databaseServerCollection{
+		aggregate: &aggregatedDatabase{
+			DatabaseServer: db,
+			status:         aggregateHealthStatuses(statuses),
+		},
+		servers: statuses,
+	}
+}
+
+type databaseServerCollection struct {
+	aggregate *aggregatedDatabase
+	servers   map[string]types.DatabaseServer
+}
+
+func (c *databaseServerCollection) get() resource {
+	return c.aggregate
+}
+
+func (c *databaseServerCollection) put(r resource) {
+	if r, ok := r.(types.DatabaseServer); ok {
+		c.servers[r.GetHostID()] = r
+		c.aggregate = &aggregatedDatabase{
+			DatabaseServer: r,
+			status:         aggregateHealthStatuses(c.servers),
+		}
+	}
+}
+
+func (c *databaseServerCollection) remove(r types.Resource) bool {
+	delete(c.servers, r.GetMetadata().Description)
+	for _, s := range c.servers {
+		c.aggregate = &aggregatedDatabase{
+			DatabaseServer: s,
+			status:         aggregateHealthStatuses(c.servers),
+		}
+		return false
+	}
+	return true
+}
+
+// aggregatedDatabase wraps a database server with aggregated health status.
+// This type exists to avoid cloning the resource unnecessarily, yet still
+// prevent data races.
+type aggregatedDatabase struct {
+	types.DatabaseServer
+	status types.TargetHealthStatus
+}
+
+func (d *aggregatedDatabase) GetTargetHealth() types.TargetHealth {
+	out := d.DatabaseServer.GetTargetHealth()
+	out.Status = string(d.status)
+	return out
+}
+
+// Copy returns a copy of the underlying database server with aggregated health
+// status.
+func (d *aggregatedDatabase) Copy() types.DatabaseServer {
+	out := d.DatabaseServer.Copy()
+	out.SetTargetHealth(d.GetTargetHealth())
+	return out
+}
+
+func (d *aggregatedDatabase) CloneResource() types.ResourceWithLabels {
+	return d.Copy()
+}
+
+func aggregateHealthStatuses[T types.TargetHealthGetter](hgs map[string]T) types.TargetHealthStatus {
+	return types.AggregateHealthStatus(func(yield func(types.TargetHealthStatus) bool) {
+		for _, hg := range hgs {
+			if !yield(types.TargetHealthStatus(hg.GetTargetHealth().Status)) {
+				return
+			}
+		}
+	})
 }
 
 type item struct {
